:sparkles: :: QR 스캔 페이지 컴포넌트로 분리
