+++ app/src/main/java/com/g3c1/oasis_android/feature_select_store/presentation/scanpage/QrScanActivity.kt
@@ -1,37 +1,17 @@
 package com.g3c1.oasis_android.feature_select_store.presentation.scanpage
 
-import android.Manifest
-import android.content.pm.PackageManager
 import android.os.Bundle
-import android.util.Size
 import androidx.activity.ComponentActivity
-import androidx.activity.compose.rememberLauncherForActivityResult
 import androidx.activity.compose.setContent
-import androidx.activity.result.contract.ActivityResultContracts
-import androidx.camera.core.CameraSelector
-import androidx.camera.core.ImageAnalysis
-import androidx.camera.core.ImageAnalysis.STRATEGY_KEEP_ONLY_LATEST
-import androidx.camera.core.Preview
-import androidx.camera.lifecycle.ProcessCameraProvider
-import androidx.camera.view.PreviewView
-import androidx.compose.foundation.Image
-import androidx.compose.foundation.background
-import androidx.compose.foundation.layout.*
-import androidx.compose.foundation.shape.RoundedCornerShape
-import androidx.compose.material.Text
-import androidx.compose.runtime.*
-import androidx.compose.ui.Alignment
+import androidx.compose.foundation.layout.Box
+import androidx.compose.foundation.layout.fillMaxSize
+import androidx.compose.runtime.getValue
+import androidx.compose.runtime.mutableStateOf
+import androidx.compose.runtime.remember
+import androidx.compose.runtime.setValue
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.draw.clip
-import androidx.compose.ui.draw.shadow
-import androidx.compose.ui.graphics.Color
-import androidx.compose.ui.platform.LocalContext
-import androidx.compose.ui.platform.LocalLifecycleOwner
-import androidx.compose.ui.res.painterResource
-import androidx.compose.ui.unit.dp
-import androidx.compose.ui.viewinterop.AndroidView
-import androidx.core.content.ContextCompat
-import com.g3c1.oasis_android.R
+import com.g3c1.oasis_android.feature_select_store.presentation.scanpage.component.QrScanScreen
+import com.g3c1.oasis_android.feature_select_store.presentation.scanpage.component.TopStoreInfoBar
 import com.g3c1.oasis_android.feature_select_store.presentation.util.QrCodeAnalyer
 
 class QrScanActivity : ComponentActivity() {
@@ -41,98 +21,14 @@ class QrScanActivity : ComponentActivity() {
             var code by remember {
                 mutableStateOf("")
             }
-            val context = LocalContext.current
-            val lifecycleOwner = LocalLifecycleOwner.current
-            val cameraProviderFuture = remember {
-                ProcessCameraProvider.getInstance(context)
-            }
-            var hasCamPermision by remember {
-                mutableStateOf(
-                    ContextCompat.checkSelfPermission(
-                        context,
-                        Manifest.permission.CAMERA
-                    ) == PackageManager.PERMISSION_GRANTED
-                )
-            }
-            val launcher = rememberLauncherForActivityResult(
-                contract = ActivityResultContracts.RequestPermission(),
-                onResult = { granted ->
-                    hasCamPermision = granted
-                }
-            )
-            LaunchedEffect(key1 = true) {
-                launcher.launch(Manifest.permission.CAMERA)
-            }
             Box(
                 modifier = Modifier
                     .fillMaxSize()
             ) {
-                Box(
-                    contentAlignment = Alignment.Center
-                ) {
-                    if (hasCamPermision) {
-                        AndroidView(
-                            factory = { context ->
-                                val previewView = PreviewView(context)
-                                val preview = Preview.Builder().build()
-                                val selector = CameraSelector.Builder()
-                                    .requireLensFacing(CameraSelector.LENS_FACING_BACK)
-                                    .build()
-                                preview.setSurfaceProvider(previewView.surfaceProvider)
-                                val imageAnalysis = ImageAnalysis.Builder()
-                                    .setTargetResolution(
-                                        Size(
-                                            previewView.width,
-                                            previewView.height
-                                        )
-                                    )
-                                    .setBackpressureStrategy(STRATEGY_KEEP_ONLY_LATEST)
-                                    .build()
-                                imageAnalysis.setAnalyzer(
-                                    ContextCompat.getMainExecutor(context),
-                                    QrCodeAnalyer { result ->
-                                        code = result
-                                    }
-                                )
-                                try {
-                                    cameraProviderFuture.get().bindToLifecycle(
-                                        lifecycleOwner,
-                                        selector,
-                                        preview,
-                                        imageAnalysis
-                                    )
-                                } catch (e: Exception) {
-                                    e.printStackTrace()
-                                }
-                                previewView
-                            },
-                            modifier = Modifier.fillMaxSize()
-                        )
-                        Image(
-                            painterResource(id = R.drawable.qr_frame),
-                            contentDescription = "QRcodeFrame"
-                        )
-                    }
-                }
-                Box(
-                    contentAlignment = Alignment.TopCenter,
-                    modifier = Modifier
-                        .fillMaxWidth()
-                        .padding(top = 10.dp)
-                ) {
-                    if (code != "") {
-                        Row(
-                            modifier = Modifier
-                                .clip(RoundedCornerShape(10.dp))
-                                .shadow(10.dp)
-                                .fillMaxWidth(0.95f)
-                                .fillMaxHeight(0.15f)
-                                .background(Color.White)
-                        ) {
-                            Text(text = code)
-                        }
-                    }
-                }
+                QrScanScreen(analyer = QrCodeAnalyer {
+                    code = it
+                })
+                TopStoreInfoBar(code = code)
             }
         }
     }
